Extract conformance runner to separate project
