+++ pkg/testing/pulumi-test-language/runner/test_host.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package conformancetestrunner
+package runner
 
 import (
 	"context"
